Fisher udpate

+++ dot_config/private_fish/functions/fisher.fish
@@ -1,6 +1,6 @@
 function fisher --argument-names cmd --description "A plugin manager for Fish"
     set --query fisher_path || set --local fisher_path $__fish_config_dir
-    set --local fisher_version 4.3.2
+    set --local fisher_version 4.3.4
     set --local fish_plugins $__fish_config_dir/fish_plugins
 
     switch "$cmd"
@@ -16,7 +16,7 @@ function fisher --argument-names cmd --description "A plugin manager for Fish"
             echo "       -v or --version  Print version"
             echo "       -h or --help     Print this help message"
             echo "Variables:"
-            echo "       \$fisher_path  Plugin installation path. Default: ~/.config/fish"
+            echo "       \$fisher_path  Plugin installation path. Default: $__fish_config_dir" | string replace --regex -- $HOME \~
         case ls list
             string match --entire --regex -- "$argv[2]" $_fisher_plugins
         case install update remove
@@ -29,13 +29,15 @@ function fisher --argument-names cmd --description "A plugin manager for Fish"
             set --local old_plugins $_fisher_plugins
             set --local new_plugins
 
+            test -e $fish_plugins && set --local file_plugins (string match --regex -- '^[^\s]+$' <$fish_plugins)
+
             if ! set --query argv[2]
                 if test "$cmd" != update
                     echo "fisher: Not enough arguments for command: \"$cmd\"" >&2 && return 1
-                else if test ! -e $fish_plugins
+                else if ! set --query file_plugins
                     echo "fisher: \"$fish_plugins\" file not found: \"$cmd\"" >&2 && return 1
                 end
-                set arg_plugins (string match --regex -- '^[^\s]+$' <$fish_plugins)
+                set arg_plugins $file_plugins
             end
 
             for plugin in $arg_plugins
@@ -70,9 +72,11 @@ function fisher --argument-names cmd --description "A plugin manager for Fish"
             set --local pid_list
             set --local source_plugins
             set --local fetch_plugins $update_plugins $install_plugins
+            set --local fish_path (status fish-path)
+
             echo (set_color --bold)fisher $cmd version $fisher_version(set_color normal)
 
-            fish_path=(status fish-path) for plugin in $fetch_plugins
+            for plugin in $fetch_plugins
                 set --local source (command mktemp -d)
                 set --append source_plugins $source
 
@@ -125,6 +129,7 @@ function fisher --argument-names cmd --description "A plugin manager for Fish"
                             emit {$name}_uninstall
                         end
                         printf "%s\n" Removing\ (set_color red --bold)$plugin(set_color normal) "         "$$plugin_files_var
+                        set --erase _fisher_plugins[$index]
                     end
 
                     command rm -rf $$plugin_files_var
@@ -134,7 +139,6 @@ function fisher --argument-names cmd --description "A plugin manager for Fish"
                         complete --erase --command $name
                     end
 
-                    set --erase _fisher_plugins[$index]
                     set --erase $plugin_files_var
                 end
             end
@@ -183,10 +187,22 @@ function fisher --argument-names cmd --description "A plugin manager for Fish"
 
             command rm -rf $source_plugins
 
-            set --query _fisher_plugins[1] || set --erase _fisher_plugins
-            set --query _fisher_plugins &&
-                printf "%s\n" $_fisher_plugins >$fish_plugins ||
+            if set --query _fisher_plugins[1]
+                set --local commit_plugins
+
+                for plugin in $file_plugins
+                    contains -- $plugin $_fisher_plugins && set --append commit_plugins $plugin
+                end
+
+                for plugin in $_fisher_plugins
+                    contains -- $plugin $commit_plugins || set --append commit_plugins $plugin
+                end
+
+                printf "%s\n" $commit_plugins >$fish_plugins
+            else
+                set --erase _fisher_plugins
                 command rm -f $fish_plugins
+            end
 
             set --local total (count $install_plugins) (count $update_plugins) (count $remove_plugins)
             test "$total" != "0 0 0" && echo (string join ", " (
